Process a list of httr2 responses

When running multiple requests in parallel using httr2::multi_req_perform() the responses are returned as a list of responses. multi_resp_process() processes the responses in a vectorized manner.

multi_resp_process(resps)

Arguments

resps

a list of httr2_response objects such as created by httr2::multi_req_perform()

Value

A list where each element is a processed FeatureCollection PBF.

Details

If a response is not 200 status code or does not have the appropriate "application/x-protobuf" content type, the result will be NULL.

The results of this are not post processed. Post processing can be applied to the resultant list object using post_process_pbf(). See the example for a fully worked example including post-processing.

See also

Examples

if (rlang::is_installed("httr2") && interactive()) {
    url <- "https://services.arcgis.com/P3ePLMYs2RVChkJx/arcgis/rest/services/ACS_Population_by_Race_and_Hispanic_Origin_Boundaries/FeatureServer/2/query?where=1=1&outFields=*&f=pbf&token="
    reqs <- replicate(5, httr2::request(url), simplify = FALSE)
    resps <- httr2::multi_req_perform(reqs)
    pbfs <- multi_resp_process(resps)
    post_process_pbf(pbfs)
}